+++ demo/demo/Demo/Distribution.hs
@@ -5,6 +5,7 @@ import Data.Word
 import Test.Tasty
 import Test.Tasty.Falsify
 
+import Test.Falsify
 import qualified Data.Falsify.WordN     as WordN
 import qualified Test.Falsify.Generator as Gen
 import qualified Test.Falsify.Range     as Range