+++ falsify/test/Main.hs
@@ -8,15 +8,6 @@ import qualified TestSuite.Sanity.Range
 import qualified TestSuite.Sanity.Selective
 import qualified TestSuite.Regression
 
-import qualified TestSuite.Prop.Generator.Compound
-import qualified TestSuite.Prop.Generator.Function
-import qualified TestSuite.Prop.Generator.Marking
-import qualified TestSuite.Prop.Generator.Precision
-import qualified TestSuite.Prop.Generator.Prim
-import qualified TestSuite.Prop.Generator.Selective
-import qualified TestSuite.Prop.Generator.Shrinking
-import qualified TestSuite.Prop.Generator.Simple
-
 main :: IO ()
 main = defaultMain $ testGroup "falsify" [
       testGroup "Sanity" [
@@ -25,15 +16,5 @@ main = defaultMain $ testGroup "falsify" [
         , TestSuite.Sanity.Predicate.tests
         ]
     , TestSuite.Regression.tests
-    , testGroup "Prop" [
-          TestSuite.Prop.Generator.Prim.tests
-        , TestSuite.Prop.Generator.Selective.tests
-        , TestSuite.Prop.Generator.Marking.tests
-        , TestSuite.Prop.Generator.Precision.tests
-        , TestSuite.Prop.Generator.Simple.tests
-        , TestSuite.Prop.Generator.Shrinking.tests
-        , TestSuite.Prop.Generator.Compound.tests
-        , TestSuite.Prop.Generator.Function.tests
-        ]
     , TestSuite.GenDefault.tests
     ]