+++ falsify/CHANGELOG.md
@@ -2,27 +2,68 @@
 
 ## 0.4.0 -- ??
 
+This release is a major cleanup release, and backwards incompatible with 0.3 in
+various ways. Below we provide a list of changes as well as a migration guide.
+
+### Package split: `falsify` vs `tasty-falsify`
+
+* The main `falsify` package no longer provides `tasty` integration (indeed,
+  it does not depend on `tasty` at all anymore); instead, `tasty` integration
+  is provided by a new `tasty-falsify` package (#80).
+
+* Module `Test.Tasty.Falsify` now lives in `tasty-falsify`, and it _only_
+  provides the functionality required for the use of `falsify` with `tasty`; it
+  no longer re-exports anything from `falsify`.
+
+* `Test.Falsify.Driver` now provides the "official" API for use in driver
+  integration.
+
+### Module hierarchy
+
+The module hierarchy has been improved.
+
+* There is a new top-level module `Test.Falsify`, which is intended to provide
+  all definitions that are meant for unqualified import. A typical `falsify`
+  test suite will therefore start with
+
+  ```hs
+  import Test.Falsify
+  import qualified Test.Falsify.Generator as Gen
+  import qualified Test.Falsify.Predicate as P
+  import qualified Test.Falsify.Range     as Range
+  ```
+
+* `Test.Falsify.Generator` now really only provides generators
+  (it was previously an awkward mix of generators and custom datatypes)
+
+* `Data.Falsify.*` is a new public module hierarchy providing some general
+  purpose data structures and utilities:
+  - `Data.Falsify.Concrete`
+  - `Data.Falsify.Permutation`
+  - `Data.Falsify.ProperFraction`
+  - `Data.Falsify.Tree`
+  - `Data.Falsify.WordN`
+
+* Some testing specific data structures now have dedicated modules:
+  - `Test.Falsify.Fun`
+  - `Test.Falsify.Marked`
+  - `Test.Falsify.SampleTree`
+  - `Test.Falsify.ShrinkTree`
+
+### Newtypes
+
+A number of type aliases have been replaced by newtypes.
+
 * In `Test.Falsify.Predicate`, `VarName` is now a newtype rather than a type
   alias, and we now also have `FnName` alongside `VarName`. Both of these have
   `IsString` instances, so code that uses `OverloadedStrings` should not be
   affected
+
 * `ShrinkTree` is now a newtype rather than a type alias, and is moved to
   `Test.Falsify.ShrinkTree`
-* Improve module hierarchy
-  - `Test.Falsify.Generator` now really only provides generators
-    (it was previously an awkward mix of generators and custom datatypes)
-  - `Data.Falsify.*` is a new public module hierarchy providing some general
-    purpose data structures and utilities:
-    - `Data.Falsify.Concrete`
-    - `Data.Falsify.Permutation`
-    - `Data.Falsify.ProperFraction`
-    - `Data.Falsify.Tree`
-    - `Data.Falsify.WordN`
-  - Some testing specific data structures now have dedicated modules:
-    - `Test.Falsify.Fun`
-    - `Test.Falsify.Marked`
-    - `Test.Falsify.SampleTree`
-    - `Test.Falsify.ShrinkTree`
+
+### Other changes
+
 * The signature of `path` has been simplified: it no longer uses `IsValidShrink`
   (which is now internal API)
 * Simplified signature of `bst`, which now only accepts inclusive bounds (#91)
@@ -36,6 +77,7 @@
   are a more natural fit when used together with `lam` to construct predicates
   of arbitrary arity. To consider the old `alwaysFail`, use `Fail "Fail"`.
 * Predicate documentation has been significantly improved.
+* Dropped support for GHC < 8.10.7
 
 ## 0.3.0 -- 2026-03-05